requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
mysql-connector-python
python-dotenv==1.0.0
flask-sqlalchemy 
mysqlclient
flask-login